+++ rdt-open-reader/src/test/java/io/ona/rdt/robolectric/shadow/UtilsShadow.java
@@ -6,7 +6,6 @@ import org.robolectric.shadow.api.Shadow;
 
 import io.ona.rdt.util.Utils;
 
-
 /**
  * Created by Vincent Karuri on 20/07/2020
  */
@@ -15,6 +14,15 @@ import io.ona.rdt.util.Utils;
 public class UtilsShadow extends Shadow {
 
     private static MockCounter mockCounter;
+    private static String parentLocationId = OpenSRPContextShadow.PARENT_LOCATION_ID;
+
+    public static String getParentLocationId() {
+        return parentLocationId;
+    }
+
+    public static void setParentLocationId(String parentLocationId) {
+        UtilsShadow.parentLocationId = parentLocationId;
+    }
 
     @Implementation
     public static void scheduleJobsImmediately() {
